Refactor device registration
